+++ PyRoute/DeltaPasses/SectorReducer.py
@@ -3,18 +3,10 @@ Created on Oct 03, 2023
 
 @author: CyberiaResurrection
 """
-from PyRoute.DeltaDebug.DeltaDictionary import DeltaDictionary
+from PyRoute.DeltaPasses.BeyondLineReducer import BeyondLineReducer
 
 
-class SectorReducer(object):
-
-    def __init__(self, reducer):
-        self.reducer = reducer
-
-    def preflight(self):
-        if self.reducer is not None and self.reducer.sectors is not None and 0 < len(self.reducer.sectors.lines):
-            return True
-        return False
+class SectorReducer(BeyondLineReducer):
 
     def run(self, singleton_only=False):
         segment = self.reducer.sectors.sector_list()
@@ -80,9 +72,3 @@ class SectorReducer(object):
         # At least one sector was shown to be irrelevant, write out the intermediate result
         if old_length > len(segment):
             self.write_files()
-
-    def write_files(self, sectors=None):
-        if isinstance(sectors, DeltaDictionary):
-            sectors.write_files(self.reducer.args.mindir)
-        else:
-            self.reducer.sectors.write_files(self.reducer.args.mindir)
